+++ src/router/dispatch/index.ts
@@ -1,3 +1,3 @@
-export * from './dispatch-by-method';
-export * from './dispatch-by-name';
-export * from './dispatch-by-pattern';
+export * from './dispatch-by.method.handler';
+export * from './dispatch-by-name.handler';
+export * from './dispatch-by-pattern.handler';
